+++ lib/src/models/response/refresh_token_response.dart
@@ -1,11 +1,11 @@
 class RefreshTokenResponse {
   final String accessToken;
-  final String refreshToken;
+  final String? refreshToken;
   final DateTime expiry;
 
   RefreshTokenResponse({
     required this.accessToken,
-    required this.refreshToken,
+    this.refreshToken,
     required this.expiry,
   });
 }