+++ lib/src/models/response/response_model.dart
@@ -1,6 +1,5 @@
-import 'package:dio_flow/src/models/response/error_type.dart';
-import 'package:dio_flow/src/models/response/links_model.dart';
-import 'package:dio_flow/src/models/response/meta_model.dart';
+import 'package:dio_flow/dio_flow.dart';
+import 'package:dio_flow/src/handlers/dio_flow_log.dart';
 
 part 'success_response_model.dart';
 part 'failed_response_model.dart';