+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.4] - 2025-09-10
+
+### 🔧 Bug Fixes
+
+- Improved debug logs
+- Optimize inputs for authentication: Remove `requiresAuth` ... if token is required: `hasBearerToken = true`
+- Add a token refresh method when receiving a 401 error: Automatic update if a `RefreshTokenHandler` is available
+
 ## [1.3.3] - 2025-09-10
 
 ### 🔧 Bug Fixes